requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --no-annotate requirements.in
#
asgiref==3.6.0
django==4.1.7
django-extensions==3.2.1
djangorestframework==3.14.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
